tangle.ct: mask redraw score without uint8 wrap-around; trust per-group decisions
residual_map returned uint8, so a group whose residual grew wrapped around to
a huge gain and every group was kept under redraw_score="mask". It now
returns int32 and the gain is taken in int64.

The --redraw-study on the Mac showed the confidence gain judges groups well
(correlation +0.73 with the truth gain; kept groups +3.76, reverted -2.10 in
total), yet whole passes were vetoed by the check that the merged fit's
total sure coverage rise. That check now only reverts a pass that loses more
than 0.005. History also logs groups_accepted.

+++ crates/tangle_python/python/tangle/ct/_confidence.py
@@ -310,11 +310,13 @@ def residual_map(
     from ._geometry import rasterize
 
     if not lines:
-        return foreground.astype(np.uint8)
+        return foreground.astype(np.int32)
     radii = np.asarray(radii, dtype=np.float64)
     labels, distance, _ = rasterize(
         foreground.shape, lines, radii, reach=radii + margin, signed=True
     )
     unexplained = foreground & (labels == 0)
     extra = ~foreground & (labels > 0) & (distance <= 0.0)
-    return (unexplained | extra).astype(np.uint8)
+    return (unexplained | extra).astype(
+        np.int32
+    )  # int32 so differences of sums cannot wrap
